Fix setup.sh

+++ setup.sh
@@ -41,7 +41,5 @@ if [ ! -d "$HOME/.config/powerline" ]; then
   ln -s $PWD/config/powerline/config.json $HOME/.config/powerline/config.json
 fi
 
-if [ ! -d "$HOME/.config/powerline/themes/tmux" ]; then
-  mkdir -p $HOME/.config/powerline/themes/tmux
-  ln -s $PWD/config/powerline/themes/tmux/default.json $HOME/.config/powerline/themes/tmux/default.json
-fi
+[ ! -d "$HOME/.config/powerline/themes/tmux" ] && mkdir -p $HOME/.config/powerline/themes/tmux
+[ ! -s "$HOME/.config/powerline/themes/tmux/default.json" ] && ln -s $PWD/config/powerline/themes/tmux/default.json $HOME/.config/powerline/themes/tmux/default.json
